Moved auto commands to auto package
